// apps/rs-marketplace/types.ts
export type PaymentStatus = 'Pago' | 'Pendente' | 'Reembolsado' | 'Parcialmente Pago' | 'Cancelado';
export type FulfillmentStatus = 'Realizado' | 'Não Realizado' | 'Parcial';

export interface ShippingAddress {
  street: string;
  number: string;
  complement?: string;
  neighborhood: string;
  city: string;
  state: string;
  zipCode: string;
}

export interface OrderItem {
  productId: string;
  variantId: string;
  productName: string;
  quantity: number;
  price: number;
  variantText?: string;
  sku?: string;
}

export interface Order {
  id: string;
  customerId?: string;
  date: string;
  customerName: string;
  customerEmail: string;
  customerCpf: string;
  customerPhone: string;
  shippingAddress: ShippingAddress;
  items: OrderItem[];
  subtotal: number;
  shippingCost: number;
  discount: number;
  appliedCoupon?: string;
  total: number;
  currency: string;
  paymentStatus: PaymentStatus;
  fulfillmentStatus: FulfillmentStatus;
  trackingCode?: string;
  shippingMethod?: string;
  notes?: string;
  boletoUrl?: string;
  pixQrCodeUrl?: string;
  pixCopyableCode?: string;
}